Make grid actions regular actions rather than auto-generated, this way they can have keyboard shortcuts

+++ zz_setContentsDialog.cpp
@@ -97,7 +97,7 @@ void SetContentsDialog::staticDialog (LDObject* obj) {
 	
 	// Mark down the history now before we perform the replacement (which
 	// destroys the old object)
-	History::addEntry (new EditHistory ({oldobj->getIndex (g_CurrentFile)},
+	History::addEntry (new EditHistory ({(ulong) oldobj->getIndex (g_CurrentFile)},
 		{oldobj->clone ()}, {obj->clone ()}));
 	
 	oldobj->replace (obj);
